ass details page added

+++ src/pages/Assignments/Assignments.jsx
@@ -4,14 +4,16 @@ import { FaCircle, FaEdit, FaEye } from 'react-icons/fa';
 import { RiDeleteBack2Fill } from "react-icons/ri";
 import { toast } from "react-toastify";
 import Swal from "sweetalert2";
+import useAuth from "../../hooks/useAuth";
+import { Link } from "react-router-dom";
 
 const Assignments = () => {
+    const { user } = useAuth();
     const [assignments, setAssignments] = useState([]);
 
     useEffect(() => {
         axios.get('http://localhost:5000/assignments')
             .then(data => {
-                console.log(data.data);
                 setAssignments(data.data)
             })
     }, []);
@@ -27,6 +29,13 @@ const Assignments = () => {
             confirmButtonText: "Yes, delete it!"
         }).then((result) => {
             if (result.isConfirmed) {
+                const assignmentToDelete = assignments.find(ass => ass._id === id);
+                console.log(assignmentToDelete?.email);
+                if (assignmentToDelete?.email !== user?.email) {
+                    toast.error("You are not authorized to delete this assignment!");
+                    return;
+                }
+
                 axios.delete(`http://localhost:5000/assignments/${id}`)
                     .then(data => {
                         console.log(data.data)
@@ -36,14 +45,17 @@ const Assignments = () => {
                                 text: "Deleted assignment successfully",
                                 icon: "success",
                                 showConfirmButton: false,
-                                timer: 1500
+                                timer: 1000
                             });
+                            const remaining = assignments.filter(ass => ass._id !== id);
+                            setAssignments(remaining);
                         }
                     })
                     .catch(error => toast.error(error.message))
             }
         });
     }
+
 
     return (
         <div className="overflow-x-auto max-w-[1400px] mx-auto">
@@ -62,7 +74,7 @@ const Assignments = () => {
                 </thead>
                 <tbody>
                     {
-                        assignments.map((assignment, i) => <tr
+                        assignments?.map((assignment, i) => <tr
                             key={assignment._id}
                             className="hover:bg-gray-100  transition duration-300 py-4">
                             <th className="py-8">{i + 1}</th>
@@ -88,9 +100,12 @@ const Assignments = () => {
                             </td>
                             <th className="py-8">
                                 <div className=" flex gap-4 items-center">
-                                    <button className="text-xl text-blue-500 hover:text-blue-500 hover:scale-110 transform transition duration-300 ease-in-out p-2 rounded-full hover:bg-blue-100 ">
-                                        <FaEdit />
-                                    </button>
+                                    <Link to={`/updateAssignment/${assignment._id}`}>
+                                        <button
+                                            className="text-xl text-blue-500 hover:text-blue-500 hover:scale-110 transform transition duration-300 ease-in-out p-2 rounded-full hover:bg-blue-100 ">
+                                            <FaEdit />
+                                        </button>
+                                    </Link>
 
                                     {/* Delete Button */}
                                     <button
@@ -100,9 +115,11 @@ const Assignments = () => {
                                     </button>
 
                                     {/* View Button */}
-                                    <button className="text-xl text-green-700 hover:text-green-500 hover:scale-110 transform transition duration-300 ease-in-out p-2 rounded-full hover:bg-green-100">
-                                        <FaEye />
-                                    </button>
+                                    <Link to={`/assignmentDetails/${assignment._id}`}>
+                                        <button className="text-xl text-green-700 hover:text-green-500 hover:scale-110 transform transition duration-300 ease-in-out p-2 rounded-full hover:bg-green-100">
+                                            <FaEye />
+                                        </button>
+                                    </Link>
                                 </div>
                             </th>
                         </tr>)
